pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitlab-mcp-bridge"
version = "0.1.0"
description = "Django-based MCP server for GitLab integration with AI-powered log analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=4.2,<6.0",
    "django-environ>=0.11.0",
    "python-gitlab>=4.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "requests>=2.31.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
]

